refactor(directive): Move inputs and outputs logic into service
So it can be reused later in other directive(s)
